+++ src/Components/Header.js
@@ -1,3 +1,9 @@
 export default function Header() {
-  return;
+  return(
+
+    <Header>
+        <h1 className="color-change-5x">RSVP App</h1>
+    </Header>
+  
+  )
 }
